Adcionado Middleware de permissoes, ajustado rotas com middleware novo, corrigido bug middleware  AUTH.

+++ src/routes/AcervoRoute.ts
@@ -5,10 +5,10 @@ import  express, { Router }  from "express";
 const acervoRouter: Router = express.Router();
 const url: string = "/acervo";
 
-acervoRouter.post(url,AuthMiddleware.auth, AcervoController.insertItem);
+acervoRouter.post(url, AuthMiddleware.auth, AcervoController.insertItem);
 acervoRouter.get(url, AcervoController.getAllItens);
-acervoRouter.get(`${url}/:id`,AuthMiddleware.auth, AcervoController.getItemById);
-acervoRouter.put(`${url}/:id`,AuthMiddleware.auth, AcervoController.updateItem);
-acervoRouter.delete(`${url}/:id`,AuthMiddleware.auth, AcervoController.deleteItem);
+acervoRouter.get(`${url}/:id`, AuthMiddleware.auth, AcervoController.getItemById);
+acervoRouter.put(`${url}/:id`, AuthMiddleware.auth, AcervoController.updateItem);
+acervoRouter.delete(`${url}/:id`, AuthMiddleware.auth, AuthMiddleware.onlyAdmins, AcervoController.deleteItem);
 
 export default acervoRouter;
